Hidden attribute (#18)
It appears that my previous PR did not work as I expected, so here's a PR that actually does what I was intending. If the previous functionality of auto-hiding the progress bar itself (but leaving its container visible) is a thing that peeps want, I can leave that in while still adding this stuff.